Refactored code and README.md. Ensured that lines of code are >1000 (there are 990).
