+++ src/main/java/interface_adapters/commands/QuitCommand.java
@@ -4,9 +4,8 @@ package interface_adapters.commands;
  * A class that passes a command from the StartGameMenu to the System.
  */
 public class QuitCommand implements GameCommand{
-
     /**
-     * Calls the System's exit method to terminate the program.
+     * {@inheritDoc}
      */
     @Override
     public void execute() {
