+++ src/main/java/use_cases/saving/HighScoreGatewayBoundary.java
@@ -1,12 +1,10 @@
 package use_cases.saving;
-
 import java.util.Map;
 
 /**
  * An interface that the FileHighScoreGateway implements to maintain the Dependency Inversion Principle.
  */
 public interface HighScoreGatewayBoundary {
-
     /**
      * Loads the highScores from file.
      * @return a Map containing the highScores
