+++ src/main/java/use_cases/scoring/EasyScoringStrategy.java
@@ -14,9 +14,7 @@ public class EasyScoringStrategy implements ScoringStrategy {
     }
 
     /**
-     * Calculates the score given the time it took to solve the anagram.
-     * @param elapsedTime a long representing how long it took
-     * @return an int representing the score
+     * {@inheritDoc}
      */
     @Override
     public int calculateScore(long elapsedTime) {
