+++ src/main/java/use_cases/scoring/HardScoringStrategy.java
@@ -4,7 +4,6 @@ package use_cases.scoring;
  * A class that implements ScoringStrategy and generates the score when the difficulty is set to hard.
  */
 public class HardScoringStrategy implements ScoringStrategy{
-
     private static int HARD_SCORE_DIVIDEND;
 
     /**
@@ -15,9 +14,7 @@ public class HardScoringStrategy implements ScoringStrategy{
     }
 
     /**
-     * Calculates the score given the time it took to solve the anagram.
-     * @param elapsedTime a long representing how long it took
-     * @return an int representing the score
+     * {@inheritDoc}
      */
     @Override
     public int calculateScore(long elapsedTime) {
